+++ NEWS.md
@@ -1,3 +1,5 @@
+# codecheck (development version)
+
 # codecheck 0.27.0
 
 ## New Features